requirements.txt
beautifulsoup4==4.12.2
Django==4.2.7
docxcompose==1.4.0
docxtpl==0.16.7
Jinja2==3.1.2
lxml==4.9.3
MarkupSafe==2.1.3
matplotlib==3.7.1
numpy==1.24.4
packaging==23.1
Pillow
python-docx==0.8.11
pytz==2023.3
cvss==2.6
django-axes
gunicorn
boto3
django-storages
dj-database-url
psycopg2-binary
whitenoise
azure-storage-blob
azure-identity